Restructuring the folders
A lot of these PCB's and simulators are for development and reference reasons only.
For simplicity I'm rearranging the folders to separate "Debugging" and the current Arduino code. I'll only be focusing on the HAN_ESP_TSS721 PCB, as the other's are either deprecated, for debugging HAN-signal or "abandoned".

This should make more sense for "end-users" who only want the code and PCB.